+++ main/local_libs/PIDController.h
@@ -1,42 +1,32 @@
 #pragma once
 namespace lyncs
 {
- class PIDController
+class PIDController
 {
   private:
-	const double kProportionGain_;
-	const double kIntegralGain_;
-	const double kDifferentialGain_;
- 	double integral_;
+	double kProportionGain_;
+	double kIntegralGain_;
+	double kDifferentialGain_;
+	double integral_;
 	double prev_error_;
 	double manipulative_variable_;
- 	PIDController();
+	PIDController();
+
   public:
 	PIDController(const double kProportionGain, const double kIntegralGain, const double kDifferentialGain);
 	~PIDController();
 	void InputPID(double data, double goal, double dt);
 	const double GetPID() const;
+	[[deprecated("Propotion gain must not be changed.")]] void SetPropotionGain(double propotion_gain) {
+		kProportionGain_ = propotion_gain;
+	}
+	[[deprecated("Integral gain must not be changed.")]] void SetIntegralGain(double integral_gain)
+	{
+		kIntegralGain_ = integral_gain;
+	}
+	[[deprecated("Differential gain must not be changed.")]] void SetDifferentialGain(double differential_gain) {
+		kDifferentialGain_ = differential_gain;
+	}
 };
- PIDController::PIDController(const double kProportionGain, const double kIntegralGain, const double kDifferentialGain)
-	: kProportionGain_(kProportionGain),
-	  kIntegralGain_(kIntegralGain),
-	  kDifferentialGain_(kDifferentialGain),
-	  integral_(0),
-	  prev_error_(0),
-	  manipulative_variable_(0)
-{
- }
- PIDController::~PIDController()
-{
-}
- void PIDController::InputPID(double data, double goal, double dt)
-{
-	double error = goal - data;
-	integral_ = error * dt;
-	manipulative_variable_ = kProportionGain_ * error ;//+ kIntegralGain_ * integral_ + kDifferentialGain_ * (error - prev_error_) / dt;
-	prev_error_ = error;
-}
- const double PIDController::GetPID() const{
-	return manipulative_variable_;
-}
- } // namespace lyncs 
+
+} // namespace lyncs
